requirements-vastai.txt

torch==2.0.0
torchvision==0.15.1
torchaudio==2.0.1
diffusers==0.17.1
safetensors==0.3.1
compel
# fairscale
# fonts
# font-roboto
numpy==1.23.5
numba==0.57.0
cupy-wheel==12.0.0
omegaconf==2.3.0
Pillow==9.5.0
pytorch_lightning
scikit-image~=0.19.3
timm~=0.6.12
einops~=0.3.2
clean-fid
resize-right
torchdiffeq
kornia
lark~=1.1.5
flask~=2.2.2
flask-socketio
bunch~=1.0.1
# waitress
click~=8.1.3
click-shell
# terminal-img
yachalk~=0.1.5
munch~=2.5.0
tqdm~=4.64.1
setuptools~=63.2.0
scipy~=1.9.3
openai-clip
# PyYAML
# packaging
requests~=2.25.1
# lpips
gdown
# facexlib
# opencv-python==4.7.0.72
opencv-contrib-python
py-jsonic
beeprint
art
gevent
gevent-websocket
websocket-client
python-socketio
transformers==4.27.4
# fire
# stumpy
python-dateutil~=2.8.2
opensimplex~=0.4.3
torchsde
# xformers==0.13.0
psutil~=5.9.4
paramiko~=2.12.0
importlib~=1.0.4
nltk~=3.8
gensim~=4.2.0
soundfile~=0.11.0
pandas~=1.5.2
pydub~=0.25.1
matplotlib~=3.6.2
stumpy~=1.11.1
basicsr~=1.4.2
fonts~=0.0.3
facexlib~=0.2.5
gfpgan~=1.3.8
realesrgan~=0.3.0
desktop-notifier
numpngw~=0.1.2
pip~=22.2.2
attrs~=22.2.0
wheel~=0.38.4
cryptography~=38.0.4
Jinja2~=3.1.2
packaging~=21.3
pyparsing~=3.0.9
# zipp~=3.11.0
Cython~=0.29.33
future~=0.18.2
mido~=1.2.10
# simpleaudio
librosa
resampy
tomesd
# git+https://github.com/facebookresearch/segment-anything.git
controlnet-aux==0.0.2
# git+https://github.com/isl-org/MiDaS.git
cupy-cuda11x
